requirements.txt
langchain
langchain-community
langchain-huggingface # using huggingface hub
huggingface_hub    # for api called LLM access 
langchain-ollama   # use this if you have ollama installed on local environment
langdetect
numpy==2.3.1
pandas==2.3.1
scikit-learn==1.7.1
joblib==1.4.2
nltk==3.9.1
streamlit
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

#second
langgraph>=0.6
langgraph-checkpoint>=2.1
langgraph-prebuilt>=0.6
langchain-core>=0.3
langsmith>=0.4
langchain-text-splitters>=0.3
# ollama>=0.5    #for ollama llms use
requests>=2.32
python-dotenv>=1.1
pyyaml>=6.0
rich>=14.0